feat: api - update analysis generation
