+++ api/resources/routes.py
@@ -3,7 +3,8 @@ from resources.view import Home
 from resources.user import ChangePassword, DeleteUser
 from resources.admin import AdminLogin, AdminLogout, ResetUserPassword, SearchUser, ActivateUser, BlockUser, UnblockedUser, AdminDeleteUser, AdminListUsers
 from resources.survey import AddSurvey, ListSurveysByUser, GetSurvey
-from resources.analysis import QueriesSurveyAnswer, GenrateDataTable
+from api.resources.analysis_datatable import GenrateDataTable
+from resources.analysis_dashboard import GetDataSummary
 
 def initialize_routes(api):
     api.add_resource(Home, '/api/home')
@@ -31,6 +32,6 @@ def initialize_routes(api):
 
     api.add_resource(AddSurvey, '/api/survey/add')
     api.add_resource(GetSurvey, '/api/survey/get/<string:survey_id>')
-    api.add_resource(QueriesSurveyAnswer, '/api/analysis/queriesurveyanswer/<string:survey_id>')
     api.add_resource(GenrateDataTable, '/api/analysis/generatedatatable/<string:survey_id>')
+    api.add_resource(GetDataSummary, '/api/analysis/getsummary/<string:survey_id>')
